LayoutTests/inspector/syntax-highlight-javascript-expected.txt

Tests that JavaScriptSourceSyntaxHighlighter detects the tokens.

return'foo';: webkit-javascript-keyword,webkit-javascript-string,*
/\//g: webkit-javascript-regexp
//ig';: webkit-javascript-comment
1 / 2 + /a/.test('a');: webkit-javascript-number,webkit-whitespace,*,webkit-whitespace,webkit-javascript-number,webkit-whitespace,*,webkit-whitespace,webkit-javascript-regexp,*,webkit-javascript-ident,webkit-brace-start,webkit-javascript-string,webkit-brace-end,*
"\"/".length / 2: webkit-javascript-string,*,webkit-javascript-ident,webkit-whitespace,*,webkit-whitespace,webkit-javascript-number
var foo = 1/*/***//2: webkit-javascript-keyword,webkit-whitespace,webkit-javascript-ident,webkit-whitespace,*,webkit-whitespace,webkit-javascript-number,webkit-javascript-comment,*,webkit-javascript-number
/*comment*//.*/.test('a'): webkit-javascript-comment,webkit-javascript-regexp,*,webkit-javascript-ident,webkit-brace-start,webkit-javascript-string,webkit-brace-end
'f\
oo';: webkit-javascript-string,,webkit-javascript-string,*
'\f\b\t';: webkit-javascript-string,*
'/\
/';: webkit-javascript-string,,webkit-javascript-string,*
foo/**
/
*/foo: webkit-javascript-ident,webkit-javascript-comment,,webkit-javascript-comment,,webkit-javascript-comment,webkit-javascript-ident
{0: true}: webkit-block-start,webkit-javascript-number,*,webkit-whitespace,webkit-javascript-keyword,webkit-block-end
var toString;: webkit-javascript-keyword,webkit-whitespace,webkit-javascript-ident,*
var foo = undefined;: webkit-javascript-keyword,webkit-whitespace,webkit-javascript-ident,webkit-whitespace,*,webkit-whitespace,webkit-javascript-undef,*
var foo = Infinity;: webkit-javascript-keyword,webkit-whitespace,webkit-javascript-ident,webkit-whitespace,*,webkit-whitespace,webkit-javascript-inf,*
var foo = NaN;: webkit-javascript-keyword,webkit-whitespace,webkit-javascript-ident,webkit-whitespace,*,webkit-whitespace,webkit-javascript-nan,*
